CustomPropertiesEditorTableItemDelegate class code moved to separate file
